Apply reviewer comments
